Multiple: Multi merge (#121)
* Tools: Add symbol converter

Add dependency to sexpdata
Add first version of symbol converter

* Ran black

* Fix comments

Remove logging artifacts
Remove static path
Readme formatting

* Add support for pin aliasing (hide)

* Improve parsing & escaping

* Add bulk conversion tool

* Add documentation for bulk conversion tool

* Increase robustness

Add better logging on parse fail
Only _warn_ about non-matched hidden pins
Add support for power symbols

* Fix kicad footprint

* Use black directly; Use click

* Finalized first prototype

Sanitize more special chars
Implement extends
Properly implement hide
Properly implement aliases
Properly implement pin names (non ints)

* WIP

* fix footprint

* implement importer; fix bugs; move test resources; improve export heuristic

* just netlist missing in sch importer

* poc kicad sch importer done

* schematic importer: import labels

* Core: IFs & CMPs better type declaration

This will make proper intellisense possible.

* Core: Remove next(); Fix IFs.unnamed

* Lib: Fix library & samples to IFs & CMPs change

* Exporter: Update faebryk prj exporter to IFs & CMPs

* Tools: Update symbol_converter to new IFs & CMPs

* Samples: Add module sample; Cleanup cmp/intfs commit

* Bugfix: Tools: symbol_converter imports

* Imports: Remove unnecessary syspath adds

* Bugfix: Tools: Symbol converter module names

* Bugfix: Mass typing fixes

* More typing

* More typing fixes

* fix last pylance errors

+++ faebryk/library/library/footprints.py
@@ -12,7 +12,7 @@ from enum import Enum
 
 
 class DIP(Footprint):
-    def __init__(self, pin_cnt: int, spacing_mm: int, long_pads: bool) -> None:
+    def __init__(self, pin_cnt: int, spacing_mm: float, long_pads: bool) -> None:
         super().__init__()
 
         class _has_kicad_footprint(has_kicad_footprint.impl()):
